Fixed issue where clicking on lines or boxes on top of an open dot would not click (or mouse over) the dot.
Implemented custom mouseOverTick so we can manage:
a) Having a large hit area for the dots on mouse over
b) Enabling mouse over on the dot when the mouse is over a line/box and
the dot.
Added custom click methods on the boxes and lines so that if the click
is directly over a dot it will register a dot click.

Added method to utils areSameDot.